Refactor healthcheck report static content into separate class.
